imports.py
# Standard library
from typing import Dict, Any, List, Optional, Tuple
from pathlib import Path
from datetime import datetime
import logging
import sys
import asyncio
import json
import hashlib
import multiprocessing

# Third-party imports
import numpy as np
import torch
import spacy
import aiohttp
import aiofiles
from langdetect import detect
import xxhash
from office365.sharepoint.client_context import ClientContext
from office365.runtime.auth.client_credential import ClientCredential

# Initialize logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)